rust fmt and cleanup (removed gaiku-2d)
