--- gaiku-2d/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "gaiku-2d"
-version = "0.1.0"
-authors = ["Norman Paniagua <normanpaniagua@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-gaiku-common = { path = "../gaiku-common", version = "0.1.0" }
