--- gaiku-2d/Readme.md
@@ -1,15 +1,0 @@
-# Gaiku 2d
-
-Implementation of the 2d terrain engine.
-
-## Features
-
-- Fixed map size
-- Infinite map size
-- Uses Quadtree to easyly query the tiles
-
-## Example
-
-```rust
-
-```
